Moved utility functions to a utils file, minor bugfix to the ruleset
